requirements.txt
streamlit
rdkit-pypi
tensorflow
pandas
numpy